SPACE-226 Geopack refactor to double (#7)
* Global refactor to double

* VICTORY!!!!

+++ src/Geopack/Geopack08/Common1.cs
@@ -3,8 +3,8 @@
 public sealed class Common1
 {
     // Elements of the rotation matrices and other parameters
-    public float ST0, CT0, SL0, CL0, CTCL, STCL, CTSL, STSL, SFI, CFI;
-    public float SPS, CPS, DS3, CGST, SGST, PSI;
-    public float A11, A21, A31, A12, A22, A32, A13, A23, A33;
-    public float E11, E21, E31, E12, E22, E32, E13, E23, E33;
+    public double ST0, CT0, SL0, CL0, CTCL, STCL, CTSL, STSL, SFI, CFI;
+    public double SPS, CPS, DS3, CGST, SGST, PSI;
+    public double A11, A21, A31, A12, A22, A32, A13, A23, A33;
+    public double E11, E21, E31, E12, E22, E32, E13, E23, E33;
 }
